requirements.txt
notebook
Pillow
numpy
